基本OID任务开发完成

+++ UserServer/src/main/java/com/ainetwork/controller/OIDController.java
@@ -7,24 +7,24 @@ import com.ainetwork.service.OIDRelationService;
 import com.ainetwork.service.OIDService;
 import com.ainetwork.service.UserOIDService;
 import com.ainetwork.util.Result;
+import com.ainetwork.vo.OIDVo;
 import com.ainetwork.vo.TreeVo;
 import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
 import com.baomidou.mybatisplus.core.metadata.IPage;
+import com.baomidou.mybatisplus.core.toolkit.StringUtils;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiOperation;
 import org.ietf.jgss.Oid;
+import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 import sun.rmi.runtime.Log;
 
 import javax.validation.constraints.NotBlank;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
-import java.util.Optional;
+import java.util.*;
 import java.util.stream.Collectors;
 
 @RestController
@@ -58,7 +58,7 @@ public class OIDController {
     @PostMapping("/register")
     public Result<?> register(@Validated  @RequestBody OID oid, Integer fatherId){
         OID byId = oidService.getById(fatherId);
-        if (Optional.ofNullable(byId).isPresent()) return Result.error("父节点不存在");
+        if (!Optional.ofNullable(byId).isPresent()) return Result.error("父节点不存在");
 
         List<Integer> sonOIDs = oidRelationService.fromFatherIdQueryOID(fatherId);
 
@@ -73,9 +73,9 @@ public class OIDController {
         OIDRelation oidRelation = new OIDRelation();
         oidRelation.setFatherId(fatherId);
 
-        oidRelation.setSonId(insert);
-
-        return oidRelationService.save(oidRelation)? Result.ok("保存oid成功"):Result.error("保存oid及其父节点失败");
+        oidRelation.setSonId(oid.getId());
+        boolean save = oidRelationService.save(oidRelation);
+        return save ? Result.ok(oid.getId()):Result.error("保存oid及其父节点失败");
     }
 
     /**
@@ -87,6 +87,8 @@ public class OIDController {
         OID byId = oidService.getById(id);
         if (byId == null) return Result.error("该oid不存在");
 
+        List<Integer> sonOIDs = oidRelationService.fromFatherIdQueryOID(id);
+        if ( sonOIDs.size() !=0 ) return Result.error("删除失败，该节点为拥有子节点的父节点，请先删除完子节点，然后才能删除父节点");
         boolean b = oidService.removeById(byId.getId());
         Boolean aBoolean = oidRelationService.deleteFromSonID(id);
         if (!b) return Result.error("删除oid失败");
@@ -102,7 +104,14 @@ public class OIDController {
 //        return Result.ok(oids);
 //    }
 
-    @ApiOperation("分页查询当前OID节点的子OID")
+    /**
+     * 带父节点
+     * @param id
+     * @param page
+     * @param pageSize
+     * @return
+     */
+    @ApiOperation("分页查询当前OID节点以及其子OID")
     @GetMapping("/userOwner")
     public Result<?> queryUserOIDPage(Integer id, Integer page, Integer pageSize){
         Page<OIDRelation> pageInfo = new Page<>(page, pageSize);
@@ -122,7 +131,7 @@ public class OIDController {
                     TreeVo treeVo1 = new TreeVo();
                     treeVo1.setId(line.getSonId());
                     OID oid  = oidService.getById(line.getSonId());
-                    treeVo1.setLabel(oid.getDescribe() + oid.getIpAddress());
+                    treeVo1.setLabel(oid.getOid()+":" + oid.getDescribe());
                     return treeVo1;
                 }
         ).collect(Collectors.toList());
@@ -130,6 +139,53 @@ public class OIDController {
         treeVo.setChildren(treeVos);
         return Result.ok(treeVo);
     }
+
+    /**
+     * 不带父节点
+     * @param id
+     * @param page
+     * @param pageSize
+     * @return
+     */
+    @ApiOperation("查询当前OID节点的子OID")
+    @GetMapping("/userOwner1")
+    public Result<?> queryUserOIDPage1(Integer id, Integer page, Integer pageSize){
+        Page<OIDRelation> pageInfo = new Page<>(page, pageSize);
+//        pageInfo  = oidService.page(pageInfo);
+        LambdaQueryWrapper<OIDRelation> wrapper = new LambdaQueryWrapper<>();
+
+        wrapper.eq(OIDRelation::getFatherId, id);
+//        OIDRelation one = oidRelationService.getOne(wrapper);
+        pageInfo = oidRelationService.page(pageInfo, wrapper);
+//        OID byId = oidService.getById(id);
+//        TreeVo treeVo = new TreeVo();
+//        treeVo.setId(byId.getId());
+//        treeVo.setLabel(byId.getDescribe()+byId.getIpAddress());
+        List<OIDRelation> records = pageInfo.getRecords();
+        if(records.size() ==0) {
+//            OIDVo oidVo = new OIDVo();
+//            oidVo.setDescribe("该节点没有子节点");
+            return Result.ok();
+        }
+        List<OIDVo> treeVos = records.stream().map(
+                line -> {
+//                    TreeVo treeVo1 = new TreeVo();
+//                    treeVo1.setId(line.getSonId());
+                    //                    treeVo1.setLabel(oid.getDescribe() + oid.getIpAddress());
+                    OID byId = oidService.getById(line.getSonId());
+                    OIDVo oidVo = new OIDVo();
+                    oidVo.setId(byId.getId());
+                    oidVo.setOid(byId.getOid());
+                    oidVo.setDescribe(byId.getDescribe());
+                    oidVo.setIpAddress(byId.getIpAddress());
+                    return oidVo;
+                }
+        ).collect(Collectors.toList());
+
+//        treeVo.setChildren(treeVos);
+        return Result.ok(treeVos);
+    }
+
 
     @ApiOperation("oid解析功能")
     @GetMapping("/oidRelation")
@@ -166,7 +222,7 @@ public class OIDController {
                 }
             }
             if(temp == null){
-                return Result.error("请检查您的oid以.分割的第"+i+1+"位置上的数，该位置上的数在数据库中没有找到");
+                return Result.error("请检查您的oid以.分割的第"+(i+1)+"位置上的数，该位置上的数在数据库中没有找到");
             }
         }
         if(i == collect.size()){
@@ -176,4 +232,70 @@ public class OIDController {
         return Result.error("oid未能成功解析");
     }
 
+    /**
+     * OID修改功能
+     */
+    @ApiOperation("OID修改功能")
+    @PostMapping("/update")
+    public Result<?> updateOID(@RequestBody OID oid){
+        if(!Optional.ofNullable(oid.getId()).isPresent()){
+            return Result.error("id不能为null");
+        }
+        OID byId = oidService.getById(oid.getId());
+        BeanUtils.copyProperties(oid, byId);
+        boolean b = oidService.updateById(byId);
+        return b ? Result.ok():Result.error();
+    }
+
+    /**
+     * 查询所有父节点OID组成的OID集合
+     */
+    @ApiOperation("查询OID完整前缀")
+    @GetMapping("/getFathers")
+    public Result<?> getFathers(Integer id){
+        List<Integer> fatherIDs = oidRelationService.queryOidAllFathter(id);
+        if (fatherIDs.size() == 0) return Result.error("oid不正确，或查询的节点为顶级OID");
+
+        StringBuilder builder = new StringBuilder();
+        for (Integer in:fatherIDs
+             ) {
+            OID byId = oidService.getById(in);
+            builder.append(byId.getOid()).append(".");
+        }
+        String string = builder.toString();
+        return Result.ok(string.substring(0, string.length()-1));
+    }
+    /**
+     * 查询当前OID下，最大的子OID数字
+     */
+    @ApiOperation("查询当前OID下，最大的子OID数字")
+    @GetMapping("/queryMaxOID")
+    public Result<?> queryMaxOID(Integer id){
+        Integer fatherIDs = oidRelationService.maxOID(id);
+        if(fatherIDs == null) return Result.ok(0);
+        OID byId = oidService.getById(fatherIDs);
+        return Result.ok(byId.getOid());
+    }
+
+    @ApiOperation("查找当前节点的父节点")
+    @GetMapping("/queryFahterOID")
+    public Result<?> getFatherOID(Integer id){
+        Integer fatherId = oidRelationService.queryFatherOID(id);
+        OID oid = oidService.getById(fatherId);
+        return Result.ok(oid);
+    }
+
+    @ApiOperation("查找当前的父节点的OID值最大的子节点")
+    @GetMapping("/queryFatherMaxOID")
+    public Result<?> getFatherMaxOID(Integer id){
+        Integer fatherId = oidRelationService.queryFatherOID(id);
+        List<Integer> sons = oidRelationService.fromFatherIdQueryOID(fatherId);
+        Integer max = 0;
+        for (Integer oidId :
+                sons) {
+            OID sonOID = oidService.getById(oidId);
+            max = max > sonOID.getOid()?max:sonOID.getOid();
+        }
+        return Result.ok(max);
+    }
 }
